requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.41
gunicorn==21.2.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
alembic==1.16.1
Mako==1.3.10
typing-extensions==4.13.2
greenlet==3.2.2
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1
humanize==4.9.0
pytz==2023.3.post1
flask-login==0.6.2
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
flask-wtf==1.1.1
marshmallow==3.19.0